refact : 변수 타입 변경
- question, reply 의 기본키를 네이밍 룰에 따라 id 를 접미사로 변경
- 엔티티 변경
- Long 타입을 int 타입으로 변경

+++ src/main/java/com/amzmall/project/payment/controller/PaymentController.java
@@ -56,7 +56,7 @@ public class PaymentController {
             @Parameter(name = "orderId", description = "상점 측 주문 고유 번호", required = true)
             @RequestParam("orderId") String orderId,
             @Parameter(name = "amount", description = "실제 결제 금액", required = true)
-            @RequestParam("amount") Long amount,
+            @RequestParam("amount") int amount,
             @Parameter(name = "paymentType", description = "결제 타입")
             @RequestParam("paymentType") PAYMENT_TYPE paymentType) {
         try {
