+++ src/main/java/com/amzmall/project/payment/domain/dto/PaymentDto.java
@@ -9,7 +9,7 @@ import lombok.*;
 @AllArgsConstructor
 public class PaymentDto {
     private String paymentType;
-    private Long amount;
+    private int amount;
     private String cardCompany;
     private String cardNumber;
     private String orderId;
